requirements.txt
alabaster==0.7.10
asn1crypto==0.22.0
Babel==2.5.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
coveralls==1.2.0
cryptography==2.0.3
Django==1.11.6
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-guardian==1.4.8
djangorestframework==3.7.3
docopt==0.6.2
docutils==0.14
flake8==3.5.0
idna==2.6
imagesize==0.7.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
model-mommy==1.5.0
mysqlclient==1.3.10
openapi-codec==1.3.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pyOpenSSL==17.2.0
PySocks==1.6.7
pytz==2017.3
PyYAML==3.12
requests==2.18.4
simplejson==3.12.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
typing==3.5.2.2
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
